Started Clean Folder
Started clean folder which contains the clean version of all the crap I had on my pi. I'm also reworking the structure of the code to be more dynamic. I'm centralizing a lot of the functionality in "FGU.py". So far have only started to set up LSM9DS1. Will eventually do the same for BMP280 and UGPS.